crons/requirements.txt
psycopg2-binary==2.9.11
requests==2.32.5